refactored demo
